스모킹 세이브 리턴값 실시간 흡연량으로 변경

+++ gbt/src/main/java/ksafinalproject/gbt/smoking/service/SmokingService.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface SmokingService {
-    int saveSmoking(ISmoking iSmoking);
+    Long saveSmoking(ISmoking iSmoking);
 
     int updateSmoking(ISmoking iSmoking, Long id);
